requirements.txt
fastapi>=0.104,<1.0
uvicorn[standard]>=0.24,<1.0
groq>=0.4,<1.0
openai>=1.0,<2.0
python-telegram-bot>=20.0,<22.0
python-dotenv>=1.0,<2.0
httpx>=0.25,<1.0
pydantic>=2.0,<3.0
